Remove most usage of python-forge and make it a doc dependency only
