+++ aiohttp_client_cache/signatures.py
@@ -13,10 +13,7 @@ Currently this is used to add the following backend-specific connection details:
 import inspect
 import re
 from logging import getLogger
-from typing import TYPE_CHECKING, Callable, Dict, Mapping, Optional, Tuple, Type, Union
-
-if TYPE_CHECKING:
-    from botocore.client import Config
+from typing import Callable, Dict, Optional, Tuple
 
 AUTOMETHOD_INIT = '.. automethod:: __init__'
 logger = getLogger(__name__)
@@ -40,47 +37,8 @@ def extend_signature(super_function: Callable, *extra_functions: Callable) -> Ca
             revision = get_combined_revision(target_function, super_function, *extra_functions)
             return revision(target_function)
         except Exception as e:
-            logger.exception(e)
+            logger.debug(e)
             return target_function
-
-    return wrapper
-
-
-def extend_init_signature(super_class: Type, *extra_functions: Callable) -> Callable:
-    """A class decorator that behaves like :py:func:`.extend_signature`, but modifies a class
-    docstring and its ``__init__`` function signature.
-    """
-
-    def wrapper(target_class: Type):
-        try:
-            # Modify init signature + docstring
-            revision = extend_signature(
-                super_class.__init__, target_class.__init__, *extra_functions
-            )
-            target_class.__init__ = revision(target_class.__init__)
-            # Include init docs in class docs
-            target_class.__doc__ = target_class.__doc__ or ''
-            if AUTOMETHOD_INIT not in target_class.__doc__:
-                target_class.__doc__ += f'\n\n    {AUTOMETHOD_INIT}\n'
-            return target_class
-        except Exception as e:
-            logger.exception(e)
-            return target_class
-
-    return wrapper
-
-
-def copy_signature(template_function: Callable, include=None, exclude=None) -> Callable:
-    """A wrapper around :py:func:`forge.copy` that silently fails if ``forge`` is not installed"""
-
-    def wrapper(target_function: Callable):
-        try:
-            import forge
-        except ImportError:
-            return target_function
-
-        revision = forge.copy(template_function, include=include, exclude=exclude)
-        return revision(target_function)
 
     return wrapper
 
@@ -175,116 +133,3 @@ def _combine_args_sections(*args_sections: str) -> str:
         args.setdefault(k.strip(), v.strip())
 
     return '\n'.join([f'    {k}: {v}' for k, v in args.items()])
-
-
-def dynamodb_template(
-    region_name: Optional[str] = None,
-    api_version: Optional[str] = None,
-    use_ssl: bool = True,
-    verify: Union[bool, str] = True,
-    endpoint_url: Optional[str] = None,
-    aws_access_key_id: Optional[str] = None,
-    aws_secret_access_key: Optional[str] = None,
-    aws_session_token: Optional[str] = None,
-    config: 'Config' = None,
-):
-    """Template function for :py:meth:`boto3.session.Session.resource`
-
-    Args:
-        region_name: The name of the region associated with the client.
-        api_version: A previous API version to use instead of the latest version
-        use_ssl: Whether or not to use SSL. Note that not all services support non-ssl connections.
-        verify: Whether or not to verify SSL certificates. You may provide either ``False`` or a path
-            to the CA cert bundle to use.
-        endpoint_url: The complete URL to use for the constructed client. If this value is provided,
-            then use_ssl is ignored.
-        aws_access_key_id : The access key to use when creating the client.
-        aws_secret_access_key: The secret key to use when creating the client.
-        aws_session_token: The session token to use when creating the client.
-        config: Advanced client configuration options. See `botocore config documentation
-            <https://botocore.amazonaws.com/v1/documentation/api/latest/reference/config.html>`_
-            for more details.
-    """
-
-
-def mongo_template(
-    host: str = 'localhost',
-    port: int = 27017,
-    document_class: Type = dict,
-    tz_aware: bool = False,
-    connect: bool = True,
-    directConnection: bool = False,
-):
-    """Template function for :py:class:`motor.motor_asyncio.AsyncIOMotorClient`
-
-    Args:
-        host: Server hostname, IP address, Unix domain socket path, or
-            `MongoDB URI <https://docs.mongodb.com/manual/reference/connection-string/>`_
-        port: Server port
-        document_class: Default class to use for documents returned from queries on this client
-        tz_aware: Return timezone-aware :py:class:`.datetime` objects
-        connect: Immediately connecting to MongoDB in the background. Otherwise connect on the
-            first operation.
-        directConnection: if True, forces this client to connect directly to the specified MongoDB
-            host as a standalone. If false, the client connects to the entire replica set of which
-            the given MongoDB host(s) is a part.
-    """
-
-
-def redis_template(
-    db: Union[str, int] = 0,
-    password: Optional[str] = None,
-    socket_timeout: Optional[float] = None,
-    socket_connect_timeout: Optional[float] = None,
-    socket_keepalive: bool = False,
-    socket_keepalive_options: Optional[Mapping[int, Union[int, bytes]]] = None,
-    socket_type: int = 0,
-    retry_on_timeout: bool = False,
-    encoding: str = "utf-8",
-    encoding_errors: str = "strict",
-    decode_responses: bool = False,
-    socket_read_size: int = 65536,
-    health_check_interval: float = 0,
-    client_name: Optional[str] = None,
-    username: Optional[str] = None,
-):
-    """Template function for :py:func:`redis.asyncio.from_url` (which passes kwargs to
-    :py:class:`redis.asyncio.Connection`)
-
-    Args:
-        db: Redis database index to switch to when connected
-        username: Username to use if Redis server instance requires authorization
-        password: Password to use if Redis server instance requires authorization
-        decode_responses: Enable response decoding
-        encoding: Codec to use for response decoding
-        socket_timeout: Timeout for a dropped connection, in seconds
-        socket_connect_timeout: Timeout for a initial connection, in seconds
-        retry_on_timeout: Retry when the connection times out
-    """
-
-
-def sqlite_template(
-    timeout: float = 5.0,
-    detect_types: int = 0,
-    isolation_level: Optional[str] = None,
-    check_same_thread: bool = True,
-    factory: Optional[Type] = None,
-    cached_statements: int = 100,
-    uri: bool = False,
-):
-    """Template function to get an accurate function signature + docs (kwargs only) for the builtin
-    :py:func:`sqlite3.connect`
-
-    Args:
-        timeout: Specifies how long the connection should wait for the lock to go away until raising
-            an exception.
-        detect_types: Can be set to any combination of ``PARSE_DECLTYPES`` and ``PARSE_COLNAMES`` to
-            turn type detection on for custom types.
-        isolation_level: Transaction isolation level. Use ``None`` for autocommit mode, or one of:
-            ``“DEFERRED”, “IMMEDIATE”, “EXCLUSIVE”``
-        check_same_thread: If True, only the creating thread may use the connection. If False, the
-            returned connection may be shared across multiple threads.
-        factory: Custom subclass of :py:class:`sqlite3.Connection` used to create connections
-        cached_statements: The number of statements that are cached internally for the connection
-        uri: Interpret database path as a URI, to allow specifying additional options
-    """
